Bump Goldberg subdivision cap from N=3 (642 cells) to N=5 (10242 cells).
The planet YAMLs already request subdivisions: 4 / 5, but solarsystem.c
was hardcoded to N=3 and the static cell buffer was sized to match.
Bumping the cap takes the per-planet vertex count over 65535 in the
worst case, so index buffers across sun/atmosphere/solarsystem
pipelines switch from uint16 to uint32 (sphere_make_uv likewise).
Scratch sized at 800k verts to absorb adversarial cliff patterns.

Per-planet subdivision still pending (single shared cell list), and
the goldberg_make() call is driven by GOLDBERG_MAX_SUBDIV until that
slice lands.

+++ src/render/sphere.h
@@ -22,4 +22,4 @@ typedef struct {
  * caller sizes the destination arrays. Returns true if everything fit. */
 bool sphere_make_uv(int stacks, int slices,
                     sphere_vertex_t *out_verts, int max_verts, int *out_v_count,
-                    uint16_t *out_indices, int max_indices, int *out_i_count);
+                    uint32_t *out_indices, int max_indices, int *out_i_count);
